backend/requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.11.3
pydantic-settings==2.9.1

# Stock data and financial analysis
yfinance==0.2.28

# AI and language models
langchain-core>=0.3.0
langchain-google-genai>=2.1.5
langgraph==0.3.31
langchain>=0.3.23
langchain-community==0.3.21

# Search and research
tavily-python==0.5.4
langchain-tavily==0.1.6
# MCP (Model Context Protocol) integration
mcp-use>=1.2.7

# Environment and utilities
python-dotenv==1.0.0

# Type hints
typing_extensions==4.13.2

# Added from the code block
polygon-api-client

# PDF generation
reportlab==4.0.7
